Fix hang on medium-sized graph import
JSNetworkX's D3 was not properly filtering the node data it was being
given, so it somehow went into an infinite loop and/or crashed when
passed nodes combined with attributes. Due to a bug with the old
workaround which broke costumes since attributes weren't actually
being merged in, a new workaround was added which added in each node
separately. However, the D3 interposition causes this to slow down,
probably due to quadratic runtime due to having to rerun layout for
every node.

This commit fixes the original JSNetworkX bug, so loading doesn't
hang Edgy when importing medium-sized graphs (~200 nodes on my
machine).