Fixed setting an attribute through adjacency
Strings can now be inputed. I also removed code that caused a string
input to delete the edge. I am unsure why deleting an edge was possible
through this block, isn’t that was set adjacency is for? If this needs
to be changed back let me know and I’ll try to get strings working with
deleting. Resolves #227

I also noticed that the get adjacency block was not working, and made a
small change to fix that.