Allow change of node shape, outline colour/size and label color
Allows the user to change the shape of a node to 'circle' or 'ellipse'
(default is 'rect').
Allows the user to change the colour of a node outline (stroke).
Allows the user to change the width of a node outline (stroke-width).
Allows the user to change the colour of a node/edge label (label-color).