Update JSNetworkX to v0.3.4
Uses JSNetworkX 0.3.4, with changes to fix the rendering of graphs.
Changes underscore jsnx names to camelCase.
Switches to the minifed version of d3.
Updates to webcola v3.
Changes internal name of PriorityQueue to PQueue to avoid conflict with
webcola.
Change default value for getting an edge attribute to 'label' instead of
'shape'